+++ .gitignore
@@ -94,3 +94,4 @@ out/
 src/Changes.txt
 .vscode/settings.json
 src/Changes.txt
+.config
