+++ config.json
@@ -1,4 +1,4 @@
 {
-  "theme": "light",
+  "theme": "dark",
   "sqliteBrowserPath": "/home/tawlib-azindo/SQLiteStudio/sqlitestudio"
 }
