--- src/main.js
@@ -1,395 +1,0 @@
-const fs = require("fs");
-const { app, BrowserWindow, ipcMain, Menu, dialog } = require("electron");
-const path = require("path");
-const { exec } = require("child_process");
-const db = require("../src/database/db.js");
-require("dotenv").config();
-
-const isDev = process.env.NODE_ENV !== "production";
-// Set up Config
-let configPath = "";
-if (isDev) {
-  configPath = path.join(__dirname, "./config.json");
-  if (!fs.existsSync(configPath)) {
-    fs.writeFileSync(configPath, JSON.stringify({}));
-  }
-} else {
-  configPath = path.join(app.getPath("userData"), "config.json");
-  if (!fs.existsSync(configPath)) {
-    fs.writeFileSync(configPath, JSON.stringify({}));
-  }
-}
-
-// load config from file system as JSON
-function loadConfig() {
-  if (!fs.existsSync(configPath)) {
-    return {
-      theme: "light",
-      sqliteBrowserPath: getDefaultSQLiteBrowserPath(),
-    };
-  }
-
-  // Load default configuration
-  return JSON.parse(fs.readFileSync(configPath, "utf8"));
-}
-
-// Save config to file system as JSON string format with 2 spaces indentation
-function saveConfig(config) {
-  fs.writeFileSync(configPath, JSON.stringify(config, null, 2), "utf8");
-}
-
-// Handle creating/removing shortcuts on Windows when installing/uninstalling.
-if (require("electron-squirrel-startup")) {
-  app.quit();
-}
-
-// setup backup for the database
-let backupDir = "";
-if (isDev) {
-  backupDir = path.join(__dirname, "./backup");
-} else {
-  backupDir = path.join(app.getPath("userData"), "backup");
-}
-// Create a backup folder if it doesn't exist
-if (!fs.existsSync(backupDir)) {
-  fs.mkdirSync(backupDir);
-}
-
-let mainWindow;
-let advanceWindow;
-// Create the browser window.
-const createMainWindow = () => {
-  mainWindow = new BrowserWindow({
-    title: "Blood Bank App",
-    width: 1200,
-    height: 800,
-    resizable: true,
-    webPreferences: {
-      nodeIntegration: true,
-      contextIsolation: true,
-      preload: path.join(__dirname, "preload.js"),
-    },
-  });
-
-  mainWindow.loadFile(path.join(__dirname, "html/index.html"));
-
-  // Create custom menu
-  const menu = Menu.buildFromTemplate(menuTemplate);
-  Menu.setApplicationMenu(menu);
-
-  const config = loadConfig();
-  applyTheme(config.theme);
-
-  // When the window is closed, close the advance window if it's open
-  mainWindow.on("closed", () => {
-    if (advanceWindow) {
-      advanceWindow.close();
-    }
-    mainWindow = null;
-  });
-};
-
-// Create the advance window
-const createAdvanceWindow = () => {
-  // Check if advance window is already open
-  if (advanceWindow) {
-    advanceWindow.focus();
-    return;
-  }
-
-  advanceWindow = new BrowserWindow({
-    title: "Advance Search",
-    width: 800,
-    height: 600,
-    frame: false,
-    // resizable: false,
-    // parent: mainWindow,
-    webPreferences: {
-      nodeIntegration: true,
-      contextIsolation: true,
-      preload: path.join(__dirname, "preload.js"),
-    },
-  });
-  advanceWindow.loadFile(path.join(__dirname, "/html/advance.html"));
-
-  // Dereference the window object when the window is closed
-  advanceWindow.on("closed", () => {
-    advanceWindow = null;
-  });
-};
-
-// Menu template
-const menuTemplate = [
-  {
-    label: "File",
-    submenu: [
-      {
-        label: "New Worksheet",
-        accelerator: "CmdOrCtrl+N",
-        click() {
-          BrowserWindow.getFocusedWindow().webContents.send(
-            "open-new-worksheet"
-          );
-        },
-      },
-      {
-        label: "General Search",
-        accelerator: "CmdOrCtrl+F",
-        click() {
-          BrowserWindow.getFocusedWindow().webContents.send(
-            "open-general-search"
-          );
-        },
-      },
-      {
-        label: "Open Database with SQLite Browser",
-        click: openSQLITEBrowser,
-      },
-      { type: "separator" },
-      {
-        label: "Exit",
-        accelerator: "CmdOrCtrl+Q",
-        role: "quit",
-      },
-    ],
-  },
-  {
-    label: "Advance",
-    submenu: [
-      {
-        label: "Advance",
-        click: createAdvanceWindow,
-      },
-    ],
-  },
-  {
-    label: "Backup",
-    submenu: [
-      {
-        label: "Backup Database",
-        click: createBackup,
-      },
-      {
-        label: "Restore Backup",
-        click: restoreBackup,
-      },
-      {
-        label: "Open Backup Folder",
-        click: openBackupFolder,
-      },
-    ],
-  },
-  {
-    label: "Settings",
-    submenu: [
-      {
-        label: "Toggle Theme",
-        click() {
-          const config = loadConfig();
-          const newTheme = config.theme === "light" ? "dark" : "light";
-          config.theme = newTheme;
-          saveConfig(config);
-          applyTheme(newTheme);
-        },
-      },
-    ],
-  },
-  {
-    label: "Help",
-    submenu: [
-      {
-        label: "Learn More",
-        click() {
-          require("electron").shell.openExternal("https://electronjs.org");
-        },
-      },
-    ],
-  },
-  ...(isDev
-    ? [
-        {
-          label: "Developer",
-          submenu: [
-            { role: "reload" },
-            { role: "forcereload" },
-            { role: "toggledevtools" },
-          ],
-        },
-      ]
-    : []),
-];
-
-// Apply theme
-function applyTheme(theme) {
-  mainWindow.webContents.send("apply-theme", theme);
-}
-
-// Backup the database
-async function createBackup() {
-  // Display the confirmation dialog
-  const result = await dialog.showMessageBox(mainWindow, {
-    type: "question",
-    buttons: ["Yes", "No"],
-    defaultId: 1,
-    title: "Backup Database",
-    message: "Are you sure you want to create a backup?",
-  });
-
-  // Check if the user clicked "Yes" or "No"
-  if (result.response === 0) {
-    const timestamp = new Date().toISOString().replace(/:/g, "-"); // For a valid file name
-    const backupPath = path.join(backupDir, `bloodBank-${timestamp}.db`);
-
-    // Copy the database file to the backup directory
-    // TODO: handle error for fs
-    fs.copyFile(db.dbPath, backupPath, async (err) => {
-      if (err) {
-        await dialog.showMessageBox(mainWindow, {
-          type: "error",
-          title: "Backup Database",
-          message: "Failed to create backup",
-        });
-      } else {
-        await dialog.showMessageBox(mainWindow, {
-          type: "info",
-          title: "Backup Database",
-          message: "Backup created successfully",
-        });
-      }
-    });
-  }
-}
-
-// Restore the database
-async function restoreBackup() {
-  // Display the file dialog
-  const result = await dialog.showOpenDialog(mainWindow, {
-    title: "Select Backup to restore",
-    defaultPath: backupDir,
-    properties: ["openFile"],
-    filters: [{ name: "Database Backups", extensions: ["db"] }],
-  });
-
-  // Check if the user selected a file
-  if (!result.canceled) {
-    const backupPath = result.filePaths[0];
-
-    // Copy the backup file to the database directory
-    fs.copyFile(backupPath, db.dbPath, async (err) => {
-      if (err) {
-        await dialog.showMessageBox(mainWindow, {
-          type: "error",
-          title: "Restore Database",
-          message: "Failed to restore backup",
-        });
-      } else {
-        await dialog.showMessageBox(mainWindow, {
-          type: "info",
-          title: "Restore Database",
-          message: "Database restored successfully",
-        });
-      }
-    });
-  }
-}
-
-// Open backup folder
-async function openBackupFolder() {
-  require("electron").shell.openPath(backupDir);
-}
-
-// Get the default path for SQLite Browser (based on platform)
-function getDefaultSQLiteBrowserPath() {
-  if (process.platform === "win32")
-    return "C:\\Program Files\\DB Browser for SQLite\\DB Browser for SQLite.exe";
-  if (process.platform === "darwin")
-    return "/Applications/DB Browser for SQLite.app/Contents/MacOS/DB Browser for SQLite";
-  return "/usr/bin/sqlitebrowser";
-}
-
-// Open database in SQLite Browser or prompt user to locate the SQLite Browser
-async function openSQLITEBrowser() {
-  const config = loadConfig();
-  let sqliteBrowserPath =
-    config.sqliteBrowserPath || getDefaultSQLiteBrowserPath();
-
-  // Check if path exists, set it if it doesn't
-  if (!fs.existsSync(sqliteBrowserPath)) {
-    const { filePaths } = await dialog.showOpenDialog(mainWindow, {
-      title: "Locate SQLite Browser",
-      properties: ["openFile"],
-    });
-    sqliteBrowserPath = filePaths[0];
-  }
-
-  exec(`"${sqliteBrowserPath}" "${dbPath}"`, (err) => {
-    if (err) {
-      dialog.showErrorBox("Error", "Failed to open database");
-    } else {
-      // Save the path to the config file
-      config.sqliteBrowserPath = sqliteBrowserPath;
-      saveConfig(config);
-    }
-  });
-}
-
-// IPC to load and save config
-ipcMain.handle("load-config", () => loadConfig());
-ipcMain.handle("save-config", (event, newConfig) => {
-  const config = { ...loadConfig(), ...newConfig };
-  saveConfig(config);
-});
-
-// IPC to handle data saving
-ipcMain.on("save-record", (event, record) => {
-  db.insertRecord(record);
-  event.returnValue = "Record saved successfully!";
-});
-
-// IPC to fetch data for a specific day
-ipcMain.on("get-records", (event, date) => {
-  const records = db.getRecords(date);
-  event.returnValue = records;
-});
-
-// IPC to fetch data for a week
-// TODO: add error handlers for invalid dates
-ipcMain.on("get-week-records", (event, startDate, endDate) => {
-  const records = db.getWeekRecords(startDate, endDate);
-  event.returnValue = records;
-});
-
-// IPC to update a record
-ipcMain.handle("update-record", async (event, updatedRecord) => {
-  db.updateRecord(updatedRecord);
-  // event.returnValue = "Record updated successfully!";
-  return true;
-});
-
-// IPC to check if a record exists for a specific date
-ipcMain.on("check-date", (event, date) => {
-  const record = db.checkDate(date);
-  // return true if a record exists for the date else false
-  event.returnValue = !!record;
-});
-
-app.whenReady().then(() => {
-  createMainWindow();
-
-  // On OS X it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  app.on("activate", () => {
-    if (BrowserWindow.getAllWindows().length === 0) {
-      createMainWindow();
-    }
-  });
-});
-
-// Quit when all windows are closed, except on macOS. There, it's common
-// for applications and their menu bar to stay active until the user quits
-// explicitly with Cmd + Q.
-app.on("window-all-closed", () => {
-  if (process.platform !== "darwin") {
-    app.quit();
-  }
-});
